% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chuuk2019}
\alias{chuuk2019}
\title{chuuk2019public}
\format{
A data frame with 75 rows and 60 variables:
\describe{
  \item{\code{Id_participant}}{double COLUMN_DESCRIPTION}
  \item{\code{Id_conflict}}{character COLUMN_DESCRIPTION}
  \item{\code{familyid}}{double COLUMN_DESCRIPTION}
  \item{\code{age}}{double COLUMN_DESCRIPTION}
  \item{\code{sex}}{character COLUMN_DESCRIPTION}
  \item{\code{conflict_causes}}{character COLUMN_DESCRIPTION}
  \item{\code{child_cry}}{double COLUMN_DESCRIPTION}
  \item{\code{suicidality_suicidal gesture}}{double COLUMN_DESCRIPTION}
  \item{\code{English_emotion_terms}}{character COLUMN_DESCRIPTION}
  \item{\code{Chuukese_emotion_terms}}{character COLUMN_DESCRIPTION}
  \item{\code{child_alliances}}{character COLUMN_DESCRIPTION}
  \item{\code{parent_alliances}}{character COLUMN_DESCRIPTION}
  \item{\code{child_intermediary}}{character COLUMN_DESCRIPTION}
  \item{\code{parent_intermediary}}{character COLUMN_DESCRIPTION}
  \item{\code{aggression}}{double COLUMN_DESCRIPTION}
  \item{\code{aggression_type}}{character COLUMN_DESCRIPTION}
  \item{\code{withhold_cooperation}}{double COLUMN_DESCRIPTION}
  \item{\code{withhold_cooperation_type}}{character COLUMN_DESCRIPTION}
  \item{\code{high_cost_strategy}}{double COLUMN_DESCRIPTION}
  \item{\code{submission}}{double COLUMN_DESCRIPTION}
  \item{\code{deception/withhold_information}}{double COLUMN_DESCRIPTION}
  \item{\code{partner_choice}}{double COLUMN_DESCRIPTION}
  \item{\code{partner_choice_who}}{character COLUMN_DESCRIPTION}
  \item{\code{negotiate_persuade_explain}}{double COLUMN_DESCRIPTION}
  \item{\code{distract}}{double COLUMN_DESCRIPTION}
  \item{\code{think_about}}{double COLUMN_DESCRIPTION}
  \item{\code{pray}}{double COLUMN_DESCRIPTION}
  \item{\code{seek_assistance}}{double COLUMN_DESCRIPTION}
  \item{\code{seek_assistance_who}}{character COLUMN_DESCRIPTION}
  \item{\code{seek_assistance_how}}{character COLUMN_DESCRIPTION}
  \item{\code{apologize}}{character COLUMN_DESCRIPTION}
  \item{\code{apology_type}}{character COLUMN_DESCRIPTION}
  \item{\code{outcome_favor}}{character COLUMN_DESCRIPTION}
  \item{\code{relationship_outcome}}{double COLUMN_DESCRIPTION}
  \item{\code{birthyear}}{double COLUMN_DESCRIPTION}
  \item{\code{adopted}}{character COLUMN_DESCRIPTION}
  \item{\code{BIOtotalsiblings}}{double COLUMN_DESCRIPTION}
  \item{\code{english}}{double COLUMN_DESCRIPTION}
  \item{\code{L_handgrip}}{double COLUMN_DESCRIPTION}
  \item{\code{R_handgrip}}{double COLUMN_DESCRIPTION}
  \item{\code{mean_handgrip}}{double COLUMN_DESCRIPTION}
  \item{\code{effort_handgrip}}{character COLUMN_DESCRIPTION}
  \item{\code{adop_parent_kin_relation}}{character COLUMN_DESCRIPTION}
  \item{\code{high_severity}}{double COLUMN_DESCRIPTION}
  \item{\code{emotion_terms_parent}}{character COLUMN_DESCRIPTION}
  \item{\code{parent_submission}}{double COLUMN_DESCRIPTION}
  \item{\code{parent_aggression}}{double COLUMN_DESCRIPTION}
  \item{\code{parent_aggression_type}}{character COLUMN_DESCRIPTION}
  \item{\code{parent_withhold_cooperation}}{double COLUMN_DESCRIPTION}
  \item{\code{parent_withhold_cooperation_type}}{character COLUMN_DESCRIPTION}
  \item{\code{parent_cry}}{double COLUMN_DESCRIPTION}
  \item{\code{parent_explain_lecture}}{double COLUMN_DESCRIPTION}
  \item{\code{parent_partner_choice}}{double COLUMN_DESCRIPTION}
  \item{\code{parent_partner_choice_who}}{character COLUMN_DESCRIPTION}
  \item{\code{which_parent_behaviors}}{character COLUMN_DESCRIPTION}
  \item{\code{location_conflict2}}{character COLUMN_DESCRIPTION}
  \item{\code{parent_conflict2}}{character COLUMN_DESCRIPTION}
  \item{\code{parent_conflict3}}{character COLUMN_DESCRIPTION}
  \item{\code{age_conflict}}{double COLUMN_DESCRIPTION}
  \item{\code{conflict_duration}}{double COLUMN_DESCRIPTION} 
}
}
\usage{
chuuk2019
}
\description{
Coded data for semi-structured interviews of Chuuk participants on parent-offspring conflict.
}
\details{
DETAILS
}
\keyword{datasets}
